color doesnot change until visibility is changed
