+++ pages/index.tsx
@@ -1,9 +1,7 @@
 import dynamic from "next/dynamic";
 
-const Map = dynamic(() => import("../views/Map"),{ssr:false});
+const Map = dynamic(() => import("../views/Map"), { ssr: false });
 
-export default function MapIndex(){
-  return(
-    <Map />
-  )
+export default function MapIndex() {
+  return <Map />;
 }
